Move library files into own sub-directory
Previously all library files lived in the root of the repository. Now
they are in a sub-directory `tuikit` and importable via:

```
import "github.com/sgeb/go-tuikit/tuikit"
```

The rationale is that I'd like to keep the "go-" prefix in the project
name, but that doesn't play well with import identifiers and as a
counter-measure I usually aliased the import.

With the new directory structure the alias is not necessary anymore and
the project is better organized to also host other files and examples.