Optimized paint request and throttling algorithm
Previously a repaint was requested through method Paint() and was
synchronous. Frame throttling was computed using time.Now(), which is a
relatively expensive call. It also led to a bug were a dropped frame
would not be redrawn if repaint wasn't explicitly requested after the
frame had been dropped.

Now tuikit.Init() returns a sending channel. The client writes to the
channel to request a repaint. This decouples painting from application
logic and allows for a more fine-grained throttling algorithm. As a
side-effect the bug mentioned earlier is fixed and there isn't any call
to time.Now() anymore.

The examples were also adjusted to use the new channel.