Remove ReadOnly from databinding properties
Having a read-only databinding property is useless after all. In order
to have a value, a read-only property would need to be set during
initialization. At that point it cannot possibly have any subscribers
yet, meaning the property's value will never propagate. In that case we
may just as well use a normal type instead of a databinding property.